+++ src/Html2Markdown/Renderers/MarkdownFormatting.cs
@@ -26,8 +26,87 @@ internal static class MarkdownFormatting
     internal static string Block(string content) =>
         $"{Environment.NewLine}{Environment.NewLine}{content}{Environment.NewLine}{Environment.NewLine}";
 
-    internal static string CollapseWhitespace(string value) =>
-        string.Join(" ", value.Split((char[]) null, StringSplitOptions.RemoveEmptyEntries));
+    internal static string CollapseWhitespace(string value)
+    {
+        var span = value.AsSpan();
+        var index = 0;
+
+        index = FindFirstNonWhiteSpaceCharacter(index, span);
+
+        if (ValueOnlyContainsWhiteSpace(index, span))
+        {
+            return string.Empty;
+        }
+
+        var builder = index > 0
+            ? new StringBuilder(value.Length)
+            : null;
+        var wordStart = index;
+        var firstWord = true;
+
+        while (index < span.Length)
+        {
+            index = FindLastNonWhiteSpaceCharacter(index, span);
+
+            var wordEnd = index;
+
+            if (builder is not null)
+            {
+                if (!firstWord)
+                {
+                    builder.Append(' ');
+                }
+
+                builder.Append(span[wordStart..wordEnd]);
+            }
+
+            if (EndOfDocumentIsReached(index, span))
+            {
+                return builder?.ToString() ?? value;
+            }
+
+            var whitespaceStart = index;
+
+            index = FindFirstNonWhiteSpaceCharacter(index, span);
+
+            var isTrailingWhitespace = index == span.Length;
+            var isSingleSpace = span[whitespaceStart] == ' ' && index - whitespaceStart == 1;
+            if (builder is null && (isTrailingWhitespace || !isSingleSpace))
+            {
+                builder = new StringBuilder(value.Length);
+                builder.Append(span[..wordEnd]);
+            }
+
+            wordStart = index;
+            firstWord = false;
+        }
+
+        return builder?.ToString() ?? value;
+    }
+
+    private static bool EndOfDocumentIsReached(int index, ReadOnlySpan<char> span) => index == span.Length;
+
+    private static int FindLastNonWhiteSpaceCharacter(int index, ReadOnlySpan<char> span)
+    {
+        while (index < span.Length && !char.IsWhiteSpace(span[index]))
+        {
+            index++;
+        }
+
+        return index;
+    }
+
+    private static bool ValueOnlyContainsWhiteSpace(int index, ReadOnlySpan<char> span) => index == span.Length;
+
+    private static int FindFirstNonWhiteSpaceCharacter(int index, ReadOnlySpan<char> span)
+    {
+        while (index < span.Length && char.IsWhiteSpace(span[index]))
+        {
+            index++;
+        }
+
+        return index;
+    }
 
     internal static string NormaliseBlockWhitespace(string markdown)
     {